+++ crates/venom-scanner/src/plugins/xxe.rs
@@ -1,7 +1,7 @@
 //! XML External Entity (XXE) Vulnerability Plugin
 
-use crate::contracts::ScanFinding;
 use crate::plugin::{Plugin, PluginCategory, PluginConfig, PluginError};
+use crate::ScanFinding;
 
 pub struct XXEPlugin;